+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Rendering: Improved performance by caching typefaces (#2833, #2848) _Thanks @KroMignon and @taya92413_
 * Avalonia: Improved performance, DPI awareness, and color rendering (#2818, #2859) _Thanks @oktrue_
 * Rename: `XMin`, `XMax`, `YMin`, and `YMax` properties are now `Left`, `Right`, `Bottom`, `Top` for all coordinate primitives (#2840)
+* Plot: Improve `AutoScale()` customization using `Margins()` to define whitespace area (#2857)
 
 ## ScottPlot 4.1.67 (in development)
 * DataLogger: Improved appearance of legend items (#2829, #2850) _Thanks @KroMignon and @p4pravin_